[Feature]: Added command handlers and Ollama integration to the bot
Detailed description of the changes:
- Defined a start() function that sends a welcome message when /start is issued
- Defined a help_command() function that sends a help message when /help is issued
- Defined a pass_to_ollama() function that echoes user messages and replies with Ollama's responses
- Added the defined command handlers to the application
- Modified the main() function to create the Application, enable logging, load bot info, and run the polling
- Created installation, restart, stop, and update batch files for easier deployment

This commit adds basic command handling functionality and Ollama integration to the Telegram bot. The bot can now respond to /start and /help commands, as well as echo user messages and provide responses using Ollama. The batch files make it easy to deploy and manage the bot on Windows systems.